requirements.txt
psycopg2-binary
aiohttp
python-dotenv

python-telegram-bot

fastapi>=0.109.1
uvicorn>=0.29.0
